debt: remove sandbox from naming
